Automate website setup
The aim of this research document is to provide guidance for arranging a business website with the use of the Vagrant and Linux platform in an automatic fashion. It also supplies parameters for the creation of the virtual machine, the installation of the web server, and sample web page deployment. This is useful while enhancing web application development environments and while configuring web servers. #Vagrant #Linux #StartupAutomation #DevOps