Automate Wordpress steup
The entire project is focused on creating a seamless WordPress installation and configuration with Vagrant and a Linux virtual machine. It employs the use of scripts to automate the installation and configuration of a LAMP stack, installations of dependencies, and deploying of WordPress. It aims to solve the problem that developers face of quickly setting up a WordPress environment for testing and development.

Major features include:

A Vagrantfile to simplify the creation and management of VMs.

Automatic generation of all components of a LAMP server (Linux, Apache, MySQL, PHP).

Scripts for generating installations and configuration of WordPress.

Alterable to suit other devop needs.